applied cr feedback
