+++ static/app/views/settings/projectDebugFiles/externalSources/customRepositories/utils.tsx
@@ -2,7 +2,14 @@ import forEach from 'lodash/forEach';
 import set from 'lodash/set';
 
 import {t} from 'app/locale';
-import {CustomRepo, CustomRepoType, customRepoTypeLabel} from 'app/types/debugFiles';
+import {CustomRepo, CustomRepoType} from 'app/types/debugFiles';
+
+export const customRepoTypeLabel = {
+  [CustomRepoType.APP_STORE_CONNECT]: 'App Store Connect',
+  [CustomRepoType.HTTP]: 'SymbolServer (HTTP)',
+  [CustomRepoType.S3]: 'Amazon S3',
+  [CustomRepoType.GCS]: 'Google Cloud Storage',
+};
 
 export const dropDownItems = [
   {
@@ -19,11 +26,6 @@ export const dropDownItems = [
     value: CustomRepoType.HTTP,
     label: customRepoTypeLabel[CustomRepoType.HTTP],
     searchKey: t('http symbol server ssqp symstore symsrv'),
-  },
-  {
-    value: CustomRepoType.APP_STORE_CONNECT,
-    label: customRepoTypeLabel[CustomRepoType.APP_STORE_CONNECT],
-    searchKey: t('apple store connect itunes ios'),
   },
 ];
